test templates splitted into separate files
